Adapted OffsetManager to support dependency injection and mocks Update kafka version Fixed some tests to avoid including EOF messages
