+++ corokafka/impl/corokafka_sent_message_impl.h
@@ -1,5 +1,5 @@
-#ifndef COROKAFKA_COROKAFKA_SENT_MESSAGE_IMPL_H
-#define COROKAFKA_COROKAFKA_SENT_MESSAGE_IMPL_H
+#ifndef BLOOMBERG_COROKAFKA_SENT_MESSAGE_IMPL_H
+#define BLOOMBERG_COROKAFKA_SENT_MESSAGE_IMPL_H
 
 #include <corokafka/interface/corokafka_isent_message.h>
 
@@ -44,4 +44,4 @@ private:
 
 }}
 
-#endif //COROKAFKA_COROKAFKA_SENT_MESSAGE_IMPL_H
+#endif //BLOOMBERG_COROKAFKA_SENT_MESSAGE_IMPL_H
