+++ corokafka/mock/corokafka_sent_message_mock.h
@@ -3,6 +3,7 @@
 
 #include <corokafka/interface/corokafka_isent_message.h>
 #include <corokafka/mock/corokafka_message_mock.h>
+#include <gmock/gmock.h>
 
 namespace Bloomberg {
 namespace corokafka {
@@ -11,6 +12,14 @@ namespace mocks {
 struct SentMessageMock : public ISentMessage,
                          public MessageMock
 {
+    explicit SentMessageMock(void* opaque = nullptr) :
+        _opaque(opaque)
+    {
+        //Set default actions
+        using namespace testing;
+        ON_CALL(*this, getOpaque())
+                .WillByDefault(Return(_opaque));
+    }
     MOCK_CONST_METHOD0(getOpaque, void*());
     
 #if (RD_KAFKA_VERSION >= RD_KAFKA_MESSAGE_STATUS_SUPPORT_VERSION)
@@ -20,6 +29,9 @@ struct SentMessageMock : public ISentMessage,
 #if RD_KAFKA_VERSION >= RD_KAFKA_MESSAGE_LATENCY_SUPPORT_VERSION
     MOCK_CONST_METHOD0(getLatency, std::chrono::microseconds());
 #endif
+
+private:
+    void* _opaque;
 };
 
 }}}
